Fix url/href issue
Previously the jist icon was added as a sibling to the a tag. then when
clicked the url that was retrieved was the 1st a tag in the parent node.
This doesn't work as links could be in the same node and might have
clicked the 2nd jist icon...
So now jist icons are added to the a tag, and thus the parent node is
the same a tag the user wants and its href is used as the url.
fixes issue on extratorrent.cc, magnet links now work again (2nd link)